fix(events): emit change events for Query without any server side config (#29130)
Review feedback: we do not want a server side config for this. Drops the
DB backed notificationSettings setting (and its Settings -> Preferences ->
Notifications page) added in the previous commit, along with the earlier
eventHandlerConfiguration.enableQueryChangeEvents yaml flag.

ChangeEventHandler dropped Entity.QUERY unconditionally, bundled in with
the Test Connection workflow exclusion. Query is now treated like every
other entity: create, update and delete produce a change event, which
consumers filter through the existing alert/webhook subscriptions. The
rest of the pipeline already expects them - ParserFactory has a
QueryFormatter and WorkflowEventConsumer lists Entity.QUERY as a valid
entity type.

Adds QueryChangeEventIT covering create, update and delete.
