+++ conf/openmetadata.yaml
@@ -346,7 +346,6 @@ database:
   checkConnectionOnBorrow: ${DB_CONNECTION_CHECK_CONNECTION_ON_BORROW:-false}  # Disable for performance
   evictionInterval: ${DB_CONNECTION_EVICTION_INTERVAL:-5 minutes}
   minIdleTime: ${DB_CONNECTION_MIN_IDLE_TIME:-5 minute}
-  queryTimeoutSeconds: ${DB_QUERY_TIMEOUT_SECONDS:-300}
 
 
   # JDBC Driver Properties - Critical for Performance
@@ -367,9 +366,9 @@ database:
     defaultRowFetchSize: ${DB_PG_DEFAULT_ROW_FETCH_SIZE:-1000}  # Fetch more rows at once
     assumeMinServerVersion: ${DB_PG_ASSUME_MIN_SERVER_VERSION:-12}  # Skip version checks
     ApplicationName: ${DB_PG_APPLICATION_NAME:-OpenMetadata}
-    loginTimeout: ${DB_PG_LOGIN_TIMEOUT:-30}  # Login timeout in seconds
-    postgresqlConnectTimeout: ${DB_POSTGRESQL_CONNECT_TIMEOUT:-30}  # Connection timeout in seconds
-    postgresqlSocketTimeout: ${DB_POSTGRESQL_SOCKET_TIMEOUT:-300}  # Socket timeout in seconds
+    loginTimeout: ${DB_PG_LOGIN_TIMEOUT:-300}  # Login timeout in seconds
+    postgresqlConnectTimeout: ${DB_POSTGRESQL_CONNECT_TIMEOUT:-60}  # Connection timeout in seconds
+    postgresqlSocketTimeout: ${DB_POSTGRESQL_SOCKET_TIMEOUT:-30000}  # Socket timeout in seconds (0 = infinite)
 
     # Aurora PostgreSQL specific optimizations
     loadBalanceHosts: ${DB_PG_LOAD_BALANCE_HOSTS:-false}  # Set to true for Aurora reader endpoints
@@ -391,7 +390,7 @@ database:
     tcpKeepAlive: ${DB_MYSQL_TCP_KEEP_ALIVE:-true}
     tcpNoDelay: ${DB_MYSQL_TCP_NO_DELAY:-true}
     mysqlConnectTimeout: ${DB_MYSQL_CONNECT_TIMEOUT:-60000}  # Connection timeout in milliseconds
-    mysqlSocketTimeout: ${DB_MYSQL_SOCKET_TIMEOUT:-300000}  # Socket timeout in milliseconds
+    mysqlSocketTimeout: ${DB_MYSQL_SOCKET_TIMEOUT:-30000000}  # Socket timeout in milliseconds (0 = infinite)
 
 objectStorage:
   enabled: ${ASSET_UPLOADER_ENABLE:-false}
@@ -796,12 +795,6 @@ bulkOperation:
   queueSize: ${BULK_OPERATION_QUEUE_SIZE:-1000}
   # Timeout in seconds for entire bulk operation
   timeoutSeconds: ${BULK_OPERATION_TIMEOUT_SECONDS:-300}
-
-# Prevent background fan-out from exhausting the request connection pool.
-asyncOperations:
-  maxConcurrentDbTasks: ${ASYNC_MAX_CONCURRENT_DB_TASKS:-25}
-  maxConcurrentRdfWrites: ${ASYNC_MAX_CONCURRENT_RDF_WRITES:-8}
-  dataInsightsMaxConcurrentDbTasks: ${DATA_INSIGHTS_MAX_CONCURRENT_DB_TASKS:-16}
 
 web:
   uriPath: ${WEB_CONF_URI_PATH:-"/api"}
