+++ openmetadata-service/src/main/java/org/openmetadata/service/events/ChangeEventHandler.java
@@ -20,14 +20,11 @@ import jakarta.ws.rs.container.ContainerResponseContext;
 import java.util.Optional;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.openmetadata.schema.configuration.NotificationSettings;
-import org.openmetadata.schema.settings.SettingsType;
 import org.openmetadata.schema.type.ChangeEvent;
 import org.openmetadata.schema.type.EventType;
 import org.openmetadata.schema.utils.JsonUtils;
 import org.openmetadata.service.Entity;
 import org.openmetadata.service.OpenMetadataApplicationConfig;
-import org.openmetadata.service.resources.settings.SettingsCache;
 import org.openmetadata.service.util.WebsocketNotificationHandler;
 
 @Slf4j
@@ -56,7 +53,8 @@ public class ChangeEventHandler implements EventHandler {
           getChangeEventFromResponseContext(responseContext, loggedInUserName);
       if (optionalChangeEvent.isPresent()) {
         ChangeEvent changeEvent = optionalChangeEvent.get();
-        if (isChangeEventSuppressed(changeEvent.getEntityType())) {
+        // Test Connection workflows shouldn't produce changeEvents (Entity.WORKFLOW)
+        if (changeEvent.getEntityType().equals(Entity.WORKFLOW)) {
           return null;
         }
         // Always set the Change Event Username as context Principal, the one creating the CE
@@ -86,27 +84,6 @@ public class ChangeEventHandler implements EventHandler {
           e);
     }
     return null;
-  }
-
-  /**
-   * Entity types that don't get recorded in the change event table. Test Connection workflows
-   * (Entity.WORKFLOW) never produce changeEvents. Queries are usually ingested in large bulk
-   * batches, so they are opt-in through Settings -> Preferences -> Notifications.
-   */
-  static boolean isChangeEventSuppressed(String entityType) {
-    if (Entity.WORKFLOW.equals(entityType)) {
-      return true;
-    }
-    return Entity.QUERY.equals(entityType) && !queryChangeEventsEnabled();
-  }
-
-  private static boolean queryChangeEventsEnabled() {
-    NotificationSettings notificationSettings =
-        SettingsCache.getSettingOrDefault(
-            SettingsType.NOTIFICATION_SETTINGS,
-            new NotificationSettings(),
-            NotificationSettings.class);
-    return Boolean.TRUE.equals(notificationSettings.getEnableQueryChangeEvents());
   }
 
   public static ChangeEvent copyChangeEvent(ChangeEvent changeEvent) {
