--- openmetadata-spec/src/main/resources/json/schema/configuration/notificationSettings.json
@@ -1,16 +1,0 @@
-{
-  "$id": "https://open-metadata.org/schema/configuration/notificationSettings.json",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "NotificationSettings",
-  "description": "Notification preferences that control which entity changes OpenMetadata records as Change Events.",
-  "type": "object",
-  "javaType": "org.openmetadata.schema.configuration.NotificationSettings",
-  "properties": {
-    "enableQueryChangeEvents": {
-      "description": "Produce Change Events for Query entity create, update, and delete operations. Disabled by default since queries are often ingested in large bulk batches.",
-      "type": "boolean",
-      "default": false
-    }
-  },
-  "additionalProperties": false
-}
