+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ar-sa.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "الاستعلامات المُعالَجة",
     "queries-scanned": "الاستعلامات المفحوصة",
     "query": "استعلام",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "مدة سجل الاستعلام",
     "query-lowercase": "استعلام",
     "query-lowercase-plural": "استعلامات",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "استيعاب البيانات الوصفية من خدمات الرسائل الأكثر استخدامًا.",
     "page-sub-header-for-metadata": "استيعاب البيانات الوصفية من خدمات البيانات الوصفية مباشرة من واجهة المستخدم.",
     "page-sub-header-for-ml-models": "استيعاب البيانات الوصفية من خدمات نماذج تعلم الآلة (ML Model) من خلال واجهة المستخدم.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "تحقق من الوصول إلى قاعدة البيانات، وصحة ES، وعميل خدمة المسار (Pipeline Service Client)، وتكوين jwks وعمليات الترحيل.",
     "page-sub-header-for-online-users": "عرض المستخدمين الذين كانوا نشطين مؤخرًا في النظام.",
     "page-sub-header-for-permission-debugger": "تصحيح وفهم أذونات المستخدم عبر الأدوار والفرق والسياسات.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "ضع عمليات بياناتك على الطيار الآلي",
     "quality-contract-description": "اختبارات وضمانات جودة البيانات",
     "queries-result-test": "ستؤدي الاستعلامات التي تُرجع صفًا واحدًا أو أكثر إلى فشل الاختبار.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "تكوين لضبط المدة التي نريد النظر إليها في سجلات الاستعلام لمعالجة بيانات الاستخدام.",
     "query-used-by-other-tables": "استعلام مستخدم من قبل جداول أخرى",
     "quick-add-tip": "استخدم الإضافة السريعة لإنشاء علاقات مصطلحات المسرد دون مغادرة الرسم البياني.",
