+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/de-de.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "Abfragen verarbeitet",
     "queries-scanned": "Abfragen gescannt",
     "query": "Abfrage",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Abfrageprotokolldauer",
     "query-lowercase": "abfrage",
     "query-lowercase-plural": "abfragen",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Ingestion von Metadaten aus den am häufigsten verwendeten Messaging-Diensten.",
     "page-sub-header-for-metadata": "Ingestion von Metadaten aus Metadatendiensten direkt über die Benutzeroberfläche.",
     "page-sub-header-for-ml-models": "Ingestion von Metadaten aus ML-Modell-Diensten über die Benutzeroberfläche.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Überprüfen Sie den Datenbankzugriff, den ES-Status, den Pipeline-Service-Client, die JWKS-Konfiguration und Migrationen.",
     "page-sub-header-for-online-users": "Benutzer anzeigen, die kürzlich im System aktiv waren.",
     "page-sub-header-for-permission-debugger": "Debuggen und verstehen Sie Benutzerberechtigungen über Rollen, Teams und Richtlinien hinweg.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Bringen Sie Ihre Datenprozesse auf Autopilot",
     "quality-contract-description": "Datenqualitätstests und -überprüfungen",
     "queries-result-test": "Abfragen, die 1 oder mehr Zeilen zurückgeben, führen zum Scheitern des Tests.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Konfiguration zur Anpassung, wie weit wir in den Abfrageprotokollen zurückblicken möchten, um Nutzungsdaten zu verarbeiten.",
     "query-used-by-other-tables": "Abfrage wird von anderen Tabellen verwendet",
     "quick-add-tip": "Verwenden Sie Schnellhinzufügen, um Glossarbegriff-Beziehungen zu erstellen, ohne den Graphen zu verlassen.",
