+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/en-us.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "queries processed",
     "queries-scanned": "queries scanned",
     "query": "Query",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Query Log Duration",
     "query-lowercase": "query",
     "query-lowercase-plural": "queries",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Ingest metadata from the most used messaging services.",
     "page-sub-header-for-metadata": "Ingest metadata from metadata services right from the UI.",
     "page-sub-header-for-ml-models": "Ingest metadata from ML Model services through the UI.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Check database access, ES health, Pipeline Service Client, jwks configuration and migrations",
     "page-sub-header-for-online-users": "View users who have been active recently in the system.",
     "page-sub-header-for-permission-debugger": "Debug and understand user permissions across roles, teams, and policies.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Put your data operations on autopilot",
     "quality-contract-description": "Data Quality Tests and Assertions",
     "queries-result-test": "Queries returning 1 or more rows will result in the test failing.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Configuration to tune how far we want to look back in query logs to process usage data.",
     "query-used-by-other-tables": "Query used by other tables",
     "quick-add-tip": "Use Quick Add to create glossary term relationships without leaving the graph.",
