+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/es-es.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "consultas procesadas",
     "queries-scanned": "consultas escaneadas",
     "query": "Consulta",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Duración del registro de consultas",
     "query-lowercase": "consulta",
     "query-lowercase-plural": "consultas",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Ingresa metadatos desde los servicios de mensajería más utilizados.",
     "page-sub-header-for-metadata": "Ingresa metadatos desde servicios de metadatos directamente desde la interfaz de usuario.",
     "page-sub-header-for-ml-models": "Ingresa metadatos desde servicios de modelo de ML a través de la interfaz de usuario.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Comprueba el acceso a base de datos, salud ES, Pipeline Service Client, configuración jwks y migraciones",
     "page-sub-header-for-online-users": "Ver usuarios que han estado activos recientemente en el sistema.",
     "page-sub-header-for-permission-debugger": "Depura y comprueba los permisos de usuario a través de roles, equipos, y políticas.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Pon tus operaciones de datos en piloto automático",
     "quality-contract-description": "Pruebas de calidad de datos y aserciones",
     "queries-result-test": "Las consultas que devuelvan 1 o más filas darán como resultado un fallo en el test.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Configuración para ajustar cuánto tiempo queremos retroceder en los registros de consultas para procesar datos de uso.",
     "query-used-by-other-tables": "Consulta utilizada por otras tablas",
     "quick-add-tip": "Use Añadir Rápido para crear relaciones entre términos del glosario sin salir del grafo.",
