+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/fr-fr.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "requêtes traitées",
     "queries-scanned": "requêtes analysées",
     "query": "Requête",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Durée du Journal de Requête",
     "query-lowercase": "requête",
     "query-lowercase-plural": "requêtes",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Ingestion de métadonnées à partir des services de messagerie les plus utilisés.",
     "page-sub-header-for-metadata": "Ingestion de métadonnées à partir des services de métadonnées directement depuis l'interface utilisateur.",
     "page-sub-header-for-ml-models": "Ingestion de métadonnées à partir des services de modèles ML via l'interface utilisateur.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Vérifiez l'accès aux bases de données, la santé d'ES, le service de Pipeline Service, la configuration jwks et les migrations",
     "page-sub-header-for-online-users": "Affichez les utilisateurs qui ont été actifs récemment dans le système.",
     "page-sub-header-for-permission-debugger": "Déboguez et comprenez les permissions des utilisateurs à travers les rôles, les équipes et les politiques.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Mettez vos opérations de données en pilote automatique",
     "quality-contract-description": "Tests et assertions de qualité des données",
     "queries-result-test": "Une requête dont le résultat retourne 1 ou plusieurs lignes entraînera l'échec du test.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Configuration pour régler la profondeur de la recherche dans les journaux de requête pour traiter les données d'utilisation.",
     "query-used-by-other-tables": "Requête utilisée par d'autres tables",
     "quick-add-tip": "Utilisez l'Ajout Rapide pour créer des relations entre termes de glossaire sans quitter le graphe.",
