+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/gl-es.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "consultas procesadas",
     "queries-scanned": "consultas escaneadas",
     "query": "Consulta",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Duración do rexistro da consulta",
     "query-lowercase": "consulta",
     "query-lowercase-plural": "consultas",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Inxesta metadatos dos servizos de mensaxería máis usados.",
     "page-sub-header-for-metadata": "Inxesta metadatos desde servizos de metadatos directamente desde a UI.",
     "page-sub-header-for-ml-models": "Inxesta metadatos dos servizos de Modelos de Aprendizaxe Automática a través da UI.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Verifica o acceso á base de datos, a saúde de Elasticsearch, o Cliente de Servizo Pipeline, a configuración de jwks e as migracións",
     "page-sub-header-for-online-users": "Ver usuarios activos recentemente no sistema.",
     "page-sub-header-for-permission-debugger": "Depura e comprende os permisos dos usuarios a través de roles, equipos e políticas.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Pon as túas operacións de datos en piloto automático",
     "quality-contract-description": "Probas de calidade de datos e asercións",
     "queries-result-test": "As consultas que devolvan 1 ou máis filas resultarán en que a proba fallará.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Configuración para axustar canto tempo queremos retroceder nos rexistros de consultas para procesar os datos de uso.",
     "query-used-by-other-tables": "Consulta usada por outras táboas",
     "quick-add-tip": "Use Engadir Rápido para crear relacións entre termos do glosario sen saír do grafo.",
