+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/he-he.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "שאילתות שעובדו",
     "queries-scanned": "שאילתות שנסרקו",
     "query": "שאילתה",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "משך לוג שאילתה",
     "query-lowercase": "שאילתה",
     "query-lowercase-plural": "שאילתות",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "שלב מטה-דאטה משירותי הזרמת המידע המובילים כגון קפקא ואחרים",
     "page-sub-header-for-metadata": "שלב מטה-דאטה משירותי מטה-דאטה אחרים הקיימים בארגון כגון Atlas ואחרים.",
     "page-sub-header-for-ml-models": "שלב מטה-דאטה ממוצרי בינה מלאכותית הקיימים בארגון.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "בדוק גישה לבסיס נתונים, בריאות ES, לקוח שירות Pipeline, תצורת jwks והגירות",
     "page-sub-header-for-online-users": "הצג משתמשים שהיו פעילים לאחרונה במערכת.",
     "page-sub-header-for-permission-debugger": "נפה באגים והבן הרשאות משתמשים בין תפקידים, צוותים ומדיניות.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "העבר את פעולות הנתונים שלך לטייס אוטומטי",
     "quality-contract-description": "בדיקות ואסרציות איכות נתונים",
     "queries-result-test": "שאילתות שמחזירות 1 או יותר שורות יכולות לגרום לכשל בבדיקה.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "הגדרות להתאמה של כמה להחזיר אחורה בלוגי שאילתות לעיבוד נתוני שימוש.",
     "query-used-by-other-tables": "שאילתה בשימוש על ידי טבלאות אחרות",
     "quick-add-tip": "השתמש בהוספה מהירה ליצירת קשרי מונחי מילון מבלי לעזוב את הגרף.",
