+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ja-jp.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "処理されたクエリ",
     "queries-scanned": "スキャンされたクエリ",
     "query": "クエリ",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "クエリログの期間",
     "query-lowercase": "クエリ",
     "query-lowercase-plural": "クエリ",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "よく使われているメッセージングサービスからメタデータを取り込みます。",
     "page-sub-header-for-metadata": "UI からメタデータサービスの情報を取り込みます。",
     "page-sub-header-for-ml-models": "UI を通じて ML モデルサービスのメタデータを取り込みます。",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "データベースアクセス、ES の状態、パイプラインクライアント、jwks 設定、マイグレーションを確認します。",
     "page-sub-header-for-online-users": "最近アクティブだったユーザーを表示します。",
     "page-sub-header-for-permission-debugger": "ロール、チーム、ポリシーをまたいでユーザー権限をデバッグ・把握します。",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "データ操作を自動運転に",
     "quality-contract-description": "データ品質テストとアサーション",
     "queries-result-test": "1件以上の行を返すクエリはテスト失敗と見なされます。",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "使用状況データを処理するために、クエリログのどこまで遡るかの設定です。",
     "query-used-by-other-tables": "他のテーブルで使用されているクエリ",
     "quick-add-tip": "クイック追加を使用すると、グラフを離れずに用語集の用語関係を作成できます。",
