+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ko-kr.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "처리된 쿼리",
     "queries-scanned": "스캔된 쿼리",
     "query": "쿼리",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "쿼리 로그 기간",
     "query-lowercase": "쿼리",
     "query-lowercase-plural": "쿼리들",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "가장 많이 사용되는 메시징 서비스에서 메타데이터를 수집하세요.",
     "page-sub-header-for-metadata": "UI에서 직접 메타데이터 서비스에서 메타데이터를 수집하세요.",
     "page-sub-header-for-ml-models": "UI를 통해 ML 모델 서비스에서 메타데이터를 수집하세요.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "데이터베이스 접근, ES 상태, 파이프라인 서비스 클라이언트, jwks 구성 및 마이그레이션을 확인하세요.",
     "page-sub-header-for-online-users": "시스템에서 최근에 활동한 사용자를 봅니다.",
     "page-sub-header-for-permission-debugger": "역할, 팀 및 정책 전반에 걸친 사용자 권한을 디버그하고 이해합니다.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "데이터 작업을 자동화하세요",
     "quality-contract-description": "데이터 품질 테스트 및 검증",
     "queries-result-test": "1개 이상의 행을 반환하는 쿼리는 테스트 실패로 결과가 나옵니다.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "사용 데이터를 처리하기 위해 쿼리 로그를 얼마나 멀리 볼지 조정하는 구성입니다.",
     "query-used-by-other-tables": "다른 테이블에서 사용된 쿼리",
     "quick-add-tip": "빠른 추가를 사용하면 그래프를 떠나지 않고 용어집 용어 관계를 만들 수 있습니다.",
