+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/mr-in.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "क्वेरी प्रक्रिया केल्या",
     "queries-scanned": "क्वेरी स्कॅन केल्या",
     "query": "क्वेरी",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "क्वेरी लॉग कालावधी",
     "query-lowercase": "क्वेरी",
     "query-lowercase-plural": "क्वेरी",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "सर्वात जास्त वापरल्या जाणार्‍या संदेशवहन सेवांमधून मेटाडेटा अंतर्ग्रहण करा.",
     "page-sub-header-for-metadata": "UI मधून मेटाडेटा सेवांमधून मेटाडेटा अंतर्ग्रहण करा.",
     "page-sub-header-for-ml-models": "UI मधून ML मॉडेल सेवांमधून मेटाडेटा अंतर्ग्रहण करा.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "डेटाबेस प्रवेश, ES आरोग्य, पाइपलाइन सेवा क्लायंट, jwks संरचना आणि स्थलांतर तपासा",
     "page-sub-header-for-online-users": "सिस्टममध्ये अलीकडे सक्रिय असलेले वापरकर्ते पहा.",
     "page-sub-header-for-permission-debugger": "भूमिका, संघ आणि धोरणांवर वापरकर्ता परवानग्या डीबग आणि समजून घ्या.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "तुमची डेटा ऑपरेशन्स ऑटोपायलटवर ठेवा",
     "quality-contract-description": "डेटा गुणवत्ता चाचण्या आणि पुष्टीकरणे",
     "queries-result-test": "1 किंवा अधिक पंक्ती परत करणाऱ्या क्वेरीमुळे चाचणी अयशस्वी होईल.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "वापर डेटा प्रक्रिया करण्यासाठी क्वेरी लॉगमध्ये मागे पाहण्याची किती दूर ट्यून करायची हे कॉन्फिगरेशन.",
     "query-used-by-other-tables": "इतर टेबल्सद्वारे वापरलेली क्वेरी",
     "quick-add-tip": "आलेख न सोडता शब्दकोश संज्ञा संबंध तयार करण्यासाठी \"जलद जोडा\" वापरा.",
