+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/nl-nl.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "query's verwerkt",
     "queries-scanned": "query's gescand",
     "query": "Query",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Duur van querylogboek",
     "query-lowercase": "query",
     "query-lowercase-plural": "query's",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Ingest metadata van de meestgebruikte berichtenservices.",
     "page-sub-header-for-metadata": "Ingest metadata van metadataservices direct vanuit de UI.",
     "page-sub-header-for-ml-models": "Ingest metadata van ML Model-services via de UI.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Controleer databasetoegang, ES-gezondheid, Pipeline Service Client, jwks-configuratie en migraties",
     "page-sub-header-for-online-users": "Bekijk gebruikers die recentelijk actief zijn geweest in het systeem.",
     "page-sub-header-for-permission-debugger": "Debug en begrijp gebruikersmachtigingen over rollen, teams en beleid.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Zet uw dataverwerking op de automatische piloot",
     "quality-contract-description": "Datakwaliteitstests en -bevestigingen",
     "queries-result-test": "Queries die 1 of meer rijen retourneren, resulteren in een mislukte test.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Configuratie om in te stellen hoe ver we willen terugkijken in querylogs om gebruiksdata te verwerken.",
     "query-used-by-other-tables": "Query gebruikt door andere tabellen",
     "quick-add-tip": "Gebruik Snel Toevoegen om glossariumtermrelaties aan te maken zonder de graf te verlaten.",
