+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/pr-pr.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "پرس‌وجوهای پردازش‌شده",
     "queries-scanned": "پرس‌وجوهای اسکن‌شده",
     "query": "پرس و جو",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "مدت زمان لاگ پرس و جو",
     "query-lowercase": "پرس و جو",
     "query-lowercase-plural": "پرس و جوها",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "ورود متادیتا از پرکاربردترین سرویس‌های پیام‌رسانی.",
     "page-sub-header-for-metadata": "ورود متادیتا از سرویس‌های متادیتا مستقیماً از رابط کاربری.",
     "page-sub-header-for-ml-models": "ورود متادیتا از سرویس‌های مدل‌های یادگیری ماشینی از طریق رابط کاربری.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "بررسی دسترسی به پایگاه داده، سلامت Elasticsearch، سرویس کلاینت Pipeline، پیکربندی jwks و مهاجرت‌ها.",
     "page-sub-header-for-online-users": "کاربرانی که اخیراً در سیستم فعال بوده‌اند را مشاهده کنید.",
     "page-sub-header-for-permission-debugger": "مجوزهای کاربر در نقش‌ها، تیم‌ها و سیاست‌ها را اشکال‌زدایی و درک کنید.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "عملیات داده‌های خود را روی خلبان خودکار بگذارید",
     "quality-contract-description": "تست‌های کیفیت داده و تأییدها",
     "queries-result-test": "پرسش‌هایی که یک یا چند ردیف را برمی‌گردانند منجر به شکست تست خواهند شد.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "پیکربندی برای تنظیم چگونگی دوربردی که می‌خواهیم به لاگ‌های پرسش نگاه کنیم تا داده‌های استفاده را پردازش کنیم.",
     "query-used-by-other-tables": "پرسش استفاده‌شده توسط جداول دیگر.",
     "quick-add-tip": "از «افزودن سریع» برای ایجاد روابط اصطلاحات واژه‌نامه بدون خروج از نمودار استفاده کنید.",
