+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/pt-br.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "consultas processadas",
     "queries-scanned": "consultas verificadas",
     "query": "Consulta",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Duração do Log de Consulta",
     "query-lowercase": "consulta",
     "query-lowercase-plural": "consultas",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Ingestão de metadados dos serviços de mensagens mais utilizados.",
     "page-sub-header-for-metadata": "Ingestão de metadados de serviços de metadados diretamente da interface do usuário.",
     "page-sub-header-for-ml-models": "Ingestão de metadados de serviços de modelo de ML por meio da interface do usuário.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Verifique o acesso ao banco de dados, o status do ES, o cliente do serviço de pipeline, a configuração do JWKS e as migrações.",
     "page-sub-header-for-online-users": "Ver usuários que estiveram ativos recentemente no sistema.",
     "page-sub-header-for-permission-debugger": "Depurar e entender permissões de usuários através de funções, times e políticas.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Coloque suas operações de dados no piloto automático",
     "quality-contract-description": "Testes de qualidade de dados e afirmações",
     "queries-result-test": "Consultas que retornam 1 ou mais linhas resultarão na falha do teste.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Configuração para ajustar até onde queremos olhar nos registros de consulta para processar dados de uso.",
     "query-used-by-other-tables": "Consulta usada por outras tabelas",
     "quick-add-tip": "Use Adição Rápida para criar relações entre termos do glossário sem sair do grafo.",
