+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ru-ru.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "запросов обработано",
     "queries-scanned": "запросов просканировано",
     "query": "Запрос",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Длительность лога запроса",
     "query-lowercase": "запрос",
     "query-lowercase-plural": "запросы",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Получение метаданных из наиболее часто используемых служб обмена сообщениями.",
     "page-sub-header-for-metadata": "Получайте метаданные из служб метаданных прямо из пользовательского интерфейса.",
     "page-sub-header-for-ml-models": "Принимайте метаданные из служб модели машинного обучения через пользовательский интерфейс.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Проверка доступа к базе данных, работоспособности ES, клиента Pipeline Service, конфигурации jwks и миграций",
     "page-sub-header-for-online-users": "Просмотр пользователей, которые недавно были активны в системе.",
     "page-sub-header-for-permission-debugger": "Отладка и понимание разрешений пользователей в ролях, командах и политиках.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Переведите операции с данными на автопилот",
     "quality-contract-description": "Тесты качества данных и утверждения",
     "queries-result-test": "Запросы, возвращающие 1 или более строк, приведут к провалу проверки.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Конфигурация того, как далеко мы хотим заглянуть в журналы логов для обработки данных об использовании.",
     "query-used-by-other-tables": "Запрос, используемый другими таблицами",
     "quick-add-tip": "Используйте «Быстрое добавление» для создания связей между терминами глоссария, не покидая граф.",
