+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/sv-se.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "frågor bearbetade",
     "queries-scanned": "frågor skannade",
     "query": "Fråga",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Varaktighet för frågelogg",
     "query-lowercase": "fråga",
     "query-lowercase-plural": "frågor",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "Mata in metadata från de mest använda meddelandetjänsterna.",
     "page-sub-header-for-metadata": "Mata in metadata från metadatatjänster direkt från gränssnittet.",
     "page-sub-header-for-ml-models": "Mata in metadata från ML Model-tjänster via gränssnittet.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Kontrollera databasåtkomst, ES-hälsa, Pipeline Service Client, jwks-konfiguration och migreringar",
     "page-sub-header-for-online-users": "Visa användare som har varit aktiva nyligen i systemet.",
     "page-sub-header-for-permission-debugger": "Felsök och förstå användarbehörigheter över roller, team och policyer.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Sätt dina dataoperationer på autopilot",
     "quality-contract-description": "Datakvalitetstester och påståenden",
     "queries-result-test": "Frågor som returnerar 1 eller fler rader gör att testet misslyckas.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Konfiguration för att finjustera hur långt tillbaka vi vill titta i frågeloggar för att bearbeta användningsdata.",
     "query-used-by-other-tables": "Fråga som används av andra tabeller",
     "quick-add-tip": "Använd Snabbtillägg för att skapa relationer mellan ordlistetermer utan att lämna grafen.",
