+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/th-th.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "เคียวรีที่ประมวลผลแล้ว",
     "queries-scanned": "เคียวรีที่สแกนแล้ว",
     "query": "การสอบถาม",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "ระยะเวลาบันทึกการสอบถาม",
     "query-lowercase": "การสอบถาม",
     "query-lowercase-plural": "การสอบถามหลายรายการ",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "นำเข้าข้อมูลเมตาจากบริการการส่งข้อความที่ใช้มากที่สุด",
     "page-sub-header-for-metadata": "นำเข้าข้อมูลเมตาจากบริการข้อมูลเมตาทางตรงจาก UI",
     "page-sub-header-for-ml-models": "นำเข้าข้อมูลเมตาจากบริการโมเดล ML ผ่าน UI",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "ตรวจสอบการเข้าถึงฐานข้อมูล, สถานะ ES, ไคลเอนต์บริการท่อ, การกำหนดค่า jwks, และการโยกย้าย",
     "page-sub-header-for-online-users": "ดูผู้ใช้ที่ใช้งานล่าสุดในระบบ",
     "page-sub-header-for-permission-debugger": "ตรวจสอบและทำความเข้าใจสิทธิ์ผู้ใช้ข้ามบทบาท ทีม และนโยบาย",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "ให้การดำเนินงานข้อมูลของคุณทำงานอัตโนมัติ",
     "quality-contract-description": "การทดสอบและการยืนยันคุณภาพข้อมูล",
     "queries-result-test": "คำสั่งที่คืนค่ามากกว่าหรือเท่ากับ 1 แถว จะทำให้การทดสอบล้มเหลว",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "การตั้งค่าเพื่อปรับแต่งระยะเวลาที่เราต้องการดูย้อนหลังในบันทึกคำสั่งเพื่อประมวลผลข้อมูลการใช้งาน",
     "query-used-by-other-tables": "คำถามที่ใช้โดยตารางอื่น",
     "quick-add-tip": "ใช้ \"เพิ่มด่วน\" เพื่อสร้างความสัมพันธ์ของคำศัพท์โดยไม่ต้องออกจากกราฟ",
