+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/tr-tr.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "işlenen sorgular",
     "queries-scanned": "taranan sorgular",
     "query": "Sorgu",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "Sorgu Günlüğü Süresi",
     "query-lowercase": "sorgu",
     "query-lowercase-plural": "sorgular",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "En çok kullanılan mesajlaşma servislerinden metadata alın.",
     "page-sub-header-for-metadata": "Metadata servislerinden doğrudan kullanıcı arayüzünden metadata alın.",
     "page-sub-header-for-ml-models": "Kullanıcı arayüzü aracılığıyla ML Model servislerinden metadata alın.",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "Veritabanı erişimini, ES sağlığını, İş Akışı Servis İstemcisini, jwks yapılandırmasını ve geçişleri kontrol edin",
     "page-sub-header-for-online-users": "Sistemde yakın zamanda aktif olan kullanıcıları görüntüleyin.",
     "page-sub-header-for-permission-debugger": "Roller, takımlar ve politikalar genelinde kullanıcı izinlerini hata ayıklayın ve anlayın.",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "Veri operasyonlarınızı otomatik pilota alın",
     "quality-contract-description": "Veri kalitesi testleri ve onaylar",
     "queries-result-test": "1 veya daha fazla satır döndüren sorgular testin başarısız olmasına neden olacaktır.",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "Kullanım verilerini işlemek için sorgu günlüklerine ne kadar geriye bakmak istediğimizi ayarlamak için yapılandırma.",
     "query-used-by-other-tables": "Diğer tablolar tarafından kullanılan sorgu",
     "quick-add-tip": "Grafikten ayrılmadan sözlük terimi ilişkileri oluşturmak için Hızlı Ekle'yi kullanın.",
