+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/zh-cn.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "已处理查询",
     "queries-scanned": "已扫描查询",
     "query": "查询",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "查询日志时长",
     "query-lowercase": "查询",
     "query-lowercase-plural": "查询",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "从最常用的消息队列类型服务中提取元数据",
     "page-sub-header-for-metadata": "通过 UI 界面, 从元数据类型服务中提取元数据",
     "page-sub-header-for-ml-models": "通过 UI 界面, 从机器学习模型类型服务中提取元数据",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "检查数据库访问、ES 健康状况、工作流服务客户端、jwks 配置和迁移情况",
     "page-sub-header-for-online-users": "查看系统中最近活跃的用户。",
     "page-sub-header-for-permission-debugger": "调试并了解跨角色、团队和策略的用户权限。",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "让您的数据操作实现自动化",
     "quality-contract-description": "数据质量测试和断言",
     "queries-result-test": "返回一行个或多行的查询将导致测试失败",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "该配置用来调整日志回溯时长, 以处理使用情况数据",
     "query-used-by-other-tables": "被其他数据表使用的查询",
     "quick-add-tip": "使用快速添加可以在不离开图的情况下创建术语表术语关系。",
