+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/zh-tw.json
@@ -2010,7 +2010,6 @@
     "queries-processed": "已處理的查詢",
     "queries-scanned": "已掃描的查詢",
     "query": "查詢",
-    "query-change-event-plural": "Query Change Events",
     "query-log-duration": "查詢日誌持續時間",
     "query-lowercase": "查詢",
     "query-lowercase-plural": "查詢",
@@ -3969,7 +3968,6 @@
     "page-sub-header-for-messagings": "從最常用的訊息服務擷取元資料。",
     "page-sub-header-for-metadata": "直接從 UI 擷取元資料服務的元資料。",
     "page-sub-header-for-ml-models": "透過 UI 從機器學習模型服務擷取元資料。",
-    "page-sub-header-for-notification-setting": "Configure the change events OpenMetadata produces for notifications and webhooks.",
     "page-sub-header-for-om-health-configuration": "檢查資料庫存取、ES 健康狀況、管線服務用戶端、jwks 組態和遷移",
     "page-sub-header-for-online-users": "檢視最近在系統中活躍的使用者。",
     "page-sub-header-for-permission-debugger": "偵錯和了解使用者在角色、團隊和政策中的權限。",
@@ -4085,7 +4083,6 @@
     "put-your-data-operations-on-autopilot": "讓您的資料操作實現自動化",
     "quality-contract-description": "資料品質測試與斷言",
     "queries-result-test": "返回 1 個或多個資料列的查詢將導致測試失敗。",
-    "query-change-event-description": "Record a change event whenever a query is created, updated, or deleted. Queries are often ingested in large bulk batches, so this is turned off by default.",
     "query-log-duration-message": "設定我們希望在查詢日誌中回溯多遠以處理使用量資料的組態。",
     "query-used-by-other-tables": "其他資料表使用的查詢",
     "quick-add-tip": "使用快速新增可在不離開圖表的情況下建立詞彙表術語關係。",
